% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon_fa.R
\name{html_dependency_fa5}
\alias{html_dependency_fa5}
\title{htmldependency for fontawesome5 in rmarkdown documents}
\usage{
html_dependency_fa5()
}
\description{
htmldependency for fontawesome5 in rmarkdown documents
}

