% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon_md.R
\name{html_dependency_md}
\alias{html_dependency_md}
\title{htmldependency for material design icons in rmarkdown documents}
\usage{
html_dependency_md()
}
\description{
htmldependency for material design icons in rmarkdown documents
}

