% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon_fa.R
\name{icon_fa}
\alias{icon_fa}
\title{Create a fontawesome 5 icon}
\usage{
icon_fa(name, iconstyle = "fas", class = NULL, style = NULL,
  transform = NULL, mask = NULL)
}
\arguments{
\item{name}{Name of icon without the fa- prefix}

\item{iconstyle}{Icon style.  By default icons are drawn as solid using the fas class.
The fab class should be included for brands and far can be used for regular,
although many icons are not available as regular in the free set}

\item{class}{Additional classes to customize the icon style such as changing the
size or animating.  See \href{https://fontawesome.com/how-to-use/svg-with-js}{usage examples}
for details.}

\item{style}{Style attributes such as color or background}

\item{transform}{Apply power transforms (position, flip and rotate) - see
{https://fontawesome.com/how-to-use/svg-with-js#power-transforms}{power transforms}
for examples}

\item{mask}{Combine two icons - see {https://fontawesome.com/how-to-use/svg-with-js#masking}{masking}
for examples}
}
\description{
Create a fontawesome 5 icon
}
\examples{
icon_fa('star', iconstyle = 'far', class = 'fa-5s')        ## large open star
icon_fa('spinner', class = 'pulse')                        ## animated spinner
icon_fa('magic', class = 'fa-5x',
    transform = 'rotate-90', style = 'color:Tomato')    ## rotated icon
\dontrun{
htmltools::div(class = 'fa-4x', icon_fa('pencil-alt',
    transform = 'shrink-10 up-.5', mask = 'fas fa-comment'))    ## mask - pencil in comment
}

}

