% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon_fa.R
\name{icon_layer}
\alias{icon_layer}
\title{layer multiple icons}
\usage{
icon_layer(icons = list())
}
\arguments{
\item{icons}{A list of font awesome 5 icons, each defined by  \code{\link{icon_fa}()}}
}
\description{
layer multiple icons
}

