% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon_md.R
\name{icon_md}
\alias{icon_md}
\title{Create a material design icon}
\usage{
icon_md(name, class = NULL)
}
\arguments{
\item{name}{Name of icon}

\item{class}{Additional classes to customize the icon style such as changing the
size.  See \href{http://google.github.io/material-design-icons/#icon-font-for-the-web}{usage examples}
for details.}
}
\description{
Create a material design icon
}
\examples{
icon_md('face')        ## face icon
icon_md('face', class = 'md-18')        ## face icon 18px
icon_md('face', class = 'md-dark md-inactive')        ## inactive, dark face icon

}

