% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon_fa.R
\name{layer_counter}
\alias{layer_counter}
\title{add a counter to an icon}
\usage{
layer_counter(counter, style = NULL)
}
\arguments{
\item{counter}{Counter value}

\item{style}{Style attributes such as color or background}
}
\description{
add a counter to an icon
}

