% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon_fa.R
\name{layer_text}
\alias{layer_text}
\title{add text to an icon}
\usage{
layer_text(counter, class = NULL, style = NULL, transform = NULL)
}
\arguments{
\item{class}{Additional classes to customize the icon style such as changing the
size or animating.  See \href{https://fontawesome.com/how-to-use/svg-with-js}{usage examples}
for details.}

\item{style}{Style attributes such as color or background}

\item{transform}{Apply power transforms (position, flip and rotate) - see
{https://fontawesome.com/how-to-use/svg-with-js#power-transforms}{power transforms}
for examples}

\item{text}{Text}
}
\description{
add text to an icon
}

